cleaned fixed colors/naming scheme
