+++ src/components/pages/constants-types.ts
@@ -5,18 +5,6 @@ export enum UserLocation {
     DASHBOARD,
 }
 
-export const renderBackground = (): string => {
-    const images = [
-        "/backgroundimage1.png",
-        "/backgroundimage2.jpg",
-        "/backgroundimage3.jpg",
-        "/backgroundimage4.jpg",
-    ];
-
-    const selector = Math.floor(Math.random() * images.length);
-    return images[selector];
-};
-
 export enum SignUpStep {
     SignUp,
     Verification,
