+++ src/components/pages/dashboard/DashboardPage.tsx
@@ -1,15 +1,10 @@
-import { useEffect, useMemo, useState } from "react";
-import { renderBackground, UserLocation } from "../constants-types"
+import { useEffect } from "react";
+import { UserLocation } from "../constants-types"
 import { NavBar } from "../NavBar"
 import { useAuth, useClerk, useSession } from "@clerk/clerk-react";
 
-import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar"
-import { AppSidebar } from "@/components/ui/app-sidebar"
-
 
 export const DashboardPage = (): JSX.Element => {
-
-    const fetchBackground = useMemo(() => renderBackground(), []);
 
     const clerk = useClerk();
     const { userId } = useAuth();
@@ -32,19 +27,12 @@ export const DashboardPage = (): JSX.Element => {
         }
     }, [isSignedIn, session])
 
-    const [theme, setTheme] = useState(
-        window.matchMedia("(prefers-color-scheme: dark)").matches ? "dark" : "light"
-    );
-
     return (
-        <SidebarProvider>
-            <AppSidebar />
-            <main className = 'flex flex-col w-full dark:bg-zinc-900 '>
-                <SidebarTrigger />
-                <AppSidebar/>
-                <div className = 'flex'>
-                </div>
-            </main>
-        </SidebarProvider>
+        <div className='dark:bg-zinc-900 dark:text-gray-100'>
+            <NavBar userLocation={UserLocation.DASHBOARD} handleSignOutConfirmClick={handleSignOutConfirmClick} />
+            <div className="pt-32 p-8 flex items-start justify-center h-screen bg-cover bg-center bg-gradient-to-b from-black via-gray-500 to-white">
+                Hello! This is the Dashboard Page.
+            </div>
+        </div>
     )
 }
