+++ src/components/pages/sign-up/SignUpPage.tsx
@@ -1,8 +1,7 @@
-import { renderBackground, SignUpStep, UserLocation } from '../constants-types'
+import { SignUpStep, UserLocation } from '../constants-types'
 import { NavBar } from '../NavBar'
 import { useEffect, useMemo, useState } from 'react';
 import { SignUpForm } from './SignUpForm';
-import { useAuth } from '@clerk/clerk-react';
 import { useNavigate } from 'react-router-dom';
 import axios from "axios";
 import { EmailVerificationForm } from './EmailVerificationForm';
@@ -36,14 +35,11 @@ export const SignUpPage = (): JSX.Element => {
         }
     }
 
-    const fetchBackground = useMemo(() => renderBackground(), []);
-
     return (
         <div className=' flex flex-col dark:bg-zinc-900 dark:text-gray-100'>
             <NavBar userLocation={UserLocation.SIGNUP} />
-            <div className="pt-20 p-8 flex flex-col items-center justify-center h-screen bg-cover bg-center bg-gradient-to-b from-pink-500 via-purple-500 to-indigo-600">
+            <div className="pt-20 p-8 flex flex-col items-center justify-center h-screen bg-cover bg-center bg-gradient-to-b from-black via-gray-500 to-white">
                 <div className="text-center z-10 relative w-full max-w-2xl mx-auto">
-                    {/* <h1 className="text-4xl font-bold p-8 text-white">Create Your Account</h1> */}
                     <div className="flex justify-center items-center">
                         {steps(currentStep)}
                     </div>
