+++ src/main.tsx
@@ -5,12 +5,11 @@ import { RouterProvider } from "react-router-dom";
 import { createBrowserRouter } from 'react-router-dom';
 
 import RootLayout from './layouts/root-layout.tsx'
-import { LandingPage } from './components/pages/LandingPage.tsx';
+import { LandingPage } from './components/pages/landing-page/LandingPage.tsx';
 import { SignUpPage } from './components/pages/sign-up/SignUpPage.tsx';
 import DashboardLayout from './layouts/dashboard-layout.tsx';
 import { DashboardPage } from './components/pages/dashboard/DashboardPage.tsx';
 import { SignInPage } from './components/pages/sign-in/SignInPage.tsx';
-import { OurPage } from './components/pages/landing-page/OurPage.tsx';
 
 const router = createBrowserRouter([
   {
@@ -19,12 +18,11 @@ const router = createBrowserRouter([
       { path: '/*', element: <LandingPage /> },
       { path: '/sign-in/*', element: <SignInPage /> },
       { path: '/sign-up/*', element: <SignUpPage /> },
-      { path: '/our-team', element: <OurPage />},
       {
         element: <DashboardLayout />,
         path: 'dashboard',
         children: [
-          { path: '/dashboard/:userId', element: <DashboardPage />},
+          { path: '/dashboard/:userId', element: <DashboardPage /> },
         ],
       },
     ],
@@ -36,4 +34,3 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
     <RouterProvider router={router} />
   </React.StrictMode>
 );
-
